# docs/openapi.json
{
  "components": {
    "schemas": {
      "AccountDetailsResponse": {
        "properties": {
          "analysis_runs": {
            "items": {
              "$ref": "#/components/schemas/AnalysisRunResponse"
            },
            "title": "Analysis Runs",
            "type": "array"
          },
          "audit_events": {
            "items": {
              "$ref": "#/components/schemas/AuditEventResponse"
            },
            "title": "Audit Events",
            "type": "array"
          },
          "clusters": {
            "title": "Clusters",
            "type": "integer"
          },
          "clusters_detail": {
            "items": {
              "$ref": "#/components/schemas/ClusterDetailResponse"
            },
            "title": "Clusters Detail",
            "type": "array"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "documents": {
            "title": "Documents",
            "type": "integer"
          },
          "documents_detail": {
            "items": {
              "$ref": "#/components/schemas/DocumentResponse"
            },
            "title": "Documents Detail",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "invitations": {
            "items": {
              "$ref": "#/components/schemas/InvitationResponse"
            },
            "title": "Invitations",
            "type": "array"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/OrganizationMembershipResponse"
            },
            "title": "Members",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "default": "active",
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "users": {
            "title": "Users",
            "type": "integer"
          },
          "workos_organization_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workos Organization Id"
          },
          "workos_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workos Updated At"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "users",
          "clusters",
          "documents"
        ],
        "title": "AccountDetailsResponse",
        "type": "object"
      },
      "AccountSummaryResponse": {
        "properties": {
          "clusters": {
            "title": "Clusters",
            "type": "integer"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "documents": {
            "title": "Documents",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "default": "active",
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "users": {
            "title": "Users",
            "type": "integer"
          },
          "workos_organization_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workos Organization Id"
          },
          "workos_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workos Updated At"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "users",
          "clusters",
          "documents"
        ],
        "title": "AccountSummaryResponse",
        "type": "object"
      },
...
      "DocumentResponse": {
        "properties": {
          "byte_size": {
            "default": 0,
            "title": "Byte Size",
            "type": "integer"
          },
          "checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "logical_document_id": {
            "title": "Logical Document Id",
            "type": "string"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "object_key": {
            "title": "Object Key",
            "type": "string"
          },
          "organization_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "published_by_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published By User Id"
          },
          "scope": {
            "description": "Document scope: organization or global.",
            "title": "Scope",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "uploaded_by_user_id": {
            "title": "Uploaded By User Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "logical_document_id",
          "version",
          "scope",
          "title",
          "filename",
          "mime_type",
          "object_key",
          "status",
          "enabled",
          "uploaded_by_user_id"
        ],
        "title": "DocumentResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CSRFHeader": {
        "description": "Signed double-submit CSRF token required for unsafe browser requests in WorkOS auth mode.",
        "in": "header",
        "name": "x-archagent-csrf",
        "type": "apiKey"
      },
      "CloudTasksOIDC": {
        "description": "Cloud Run OIDC token presented by Google Cloud Tasks in production.",
        "scheme": "bearer",
        "type": "http"
      },
      "CollectorBearer": {
        "description": "Collector credential returned by /collector/v1/register or /collector/v1/credentials/rotate.",
        "scheme": "bearer",
        "type": "http"
      },
      "DevelopmentIdentityHeader": {
        "description": "Local development/test identity selector. Disabled in production.",
        "in": "header",
        "name": "x-archagent-user",
        "type": "apiKey"
      },
      "WorkOSSessionCookie": {
        "description": "Sealed WorkOS AuthKit session cookie used in hosted auth mode.",
        "in": "cookie",
        "name": "__Host-archagent-session",
        "type": "apiKey"
      },
      "WorkOSWebhookSignature": {
        "description": "WorkOS webhook signature header validated with the configured webhook secret.",
        "in": "header",
        "name": "workos-signature",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "\nArchAgent API contract for browser frontend, collector, staff, webhook, and worker integrations.\n\nUse `/openapi.json` as the machine-readable source of truth. Swagger UI is available at `/docs`\nand ReDoc is available at `/redoc` when the FastAPI app is running.\n",
    "title": "ArchAgent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/v1/accounts": {
      "get": {
        "description": "Return staff-only account summaries with user, cluster, and document counts.",
        "operationId": "list_accounts_internal_v1_accounts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccountSummaryResponse"
                  },
                  "title": "Response List Accounts Internal V1 Accounts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "List Accounts",
        "tags": [
          "Internal"
        ]
      }
    },
    "/internal/v1/accounts/{account_id}": {
      "get": {
        "description": "Return one staff-only account detail payload with members, invitations, clusters, documents, analysis runs, and audit events.",
        "operationId": "get_account_internal_v1_accounts__account_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDetailsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "Get Account Details",
        "tags": [
          "Internal"
        ]
      }
    },
    "/internal/v1/knowledge/documents": {
      "get": {
        "description": "Return staff-only global knowledge documents.",
        "operationId": "list_global_documents_internal_v1_knowledge_documents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  },
                  "title": "Response List Global Documents Internal V1 Knowledge Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "List Global Knowledge Documents",
        "tags": [
          "Internal"
        ]
      }
    },
    "/internal/v1/knowledge/documents/uploads": {
      "post": {
        "description": "Create global document metadata and return the API-relative URL for uploading bytes.",
        "operationId": "create_global_upload_internal_v1_knowledge_documents_uploads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalDocumentUploadCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "CSRFHeader": [],
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "Create Global Knowledge Upload",
        "tags": [
          "Internal"
        ]
      }
    },
    "/internal/v1/knowledge/documents/{document_id}/archive": {
      "post": {
        "description": "Archive and disable a published global knowledge document.",
        "operationId": "archive_global_document_internal_v1_knowledge_documents__document_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "CSRFHeader": [],
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "Archive Global Knowledge Document",
        "tags": [
          "Internal"
        ]
      }
    },
    "/internal/v1/knowledge/documents/{document_id}/complete": {
      "post": {
        "description": "Mark a global document ready for asynchronous scanning, chunking, and indexing.",
        "operationId": "complete_global_upload_internal_v1_knowledge_documents__document_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "CSRFHeader": [],
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "Complete Global Knowledge Upload",
        "tags": [
          "Internal"
        ]
      }
    },
    "/internal/v1/knowledge/documents/{document_id}/content": {
      "put": {
        "description": "Upload raw bytes for a global knowledge document.",
        "operationId": "upload_global_content_internal_v1_knowledge_documents__document_id__content_put",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "CSRFHeader": [],
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "Upload Global Knowledge Content",
        "tags": [
          "Internal"
        ]
      }
    },
    "/internal/v1/knowledge/documents/{document_id}/publish": {
      "post": {
        "description": "Publish an indexed draft or archived global document for retrieval across organizations.",
        "operationId": "publish_global_document_internal_v1_knowledge_documents__document_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "CSRFHeader": [],
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "Publish Global Knowledge Document",
        "tags": [
          "Internal"
        ]
      }
    },
    "/internal/v1/knowledge/documents/{document_id}/rollback": {
      "post": {
        "description": "Republish a previous global document version.",
        "operationId": "rollback_global_document_internal_v1_knowledge_documents__document_id__rollback_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "CSRFHeader": [],
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "Rollback Global Knowledge Document",
        "tags": [
          "Internal"
        ]
      }
    },
    "/internal/v1/users": {
      "get": {
        "description": "Return staff-only user records with organization memberships.",
        "operationId": "list_users_internal_v1_users_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  },
                  "title": "Response List Users Internal V1 Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "List Users",
        "tags": [
          "Internal"
        ]
      }
    },
...
    "/v1/knowledge/documents": {
      "get": {
        "description": "List organization-scoped knowledge documents available to the active organization.",
        "operationId": "get_documents_v1_knowledge_documents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  },
                  "title": "Response Get Documents V1 Knowledge Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "List Knowledge Documents",
        "tags": [
          "Product"
        ]
      }
    },
    "/v1/knowledge/documents/uploads": {
      "post": {
        "description": "Create document metadata and return the API-relative URL for uploading bytes.",
        "operationId": "create_document_upload_v1_knowledge_documents_uploads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentUploadCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "CSRFHeader": [],
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "Create Knowledge Document Upload",
        "tags": [
          "Product"
        ]
      }
    },
...
    "/v1/knowledge/documents/{document_id}/complete": {
      "post": {
        "description": "Mark an uploaded document ready for asynchronous scanning, chunking, and indexing.",
        "operationId": "complete_document_upload_v1_knowledge_documents__document_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "CSRFHeader": [],
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "Complete Knowledge Document Upload",
        "tags": [
          "Product"
        ]
      }
    },
    "/v1/knowledge/documents/{document_id}/content": {
      "put": {
        "description": "Upload raw document bytes to the URL returned by the upload creation endpoint.",
        "operationId": "upload_document_content_v1_knowledge_documents__document_id__content_put",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "CSRFHeader": [],
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "Upload Knowledge Document Content",
        "tags": [
          "Product"
        ]
      }
    },
...
  },
...
}